set all components separately
